More code re-org.
